Refactor map components and update database connection imports
