+++ src/app/dashboard/map/Map Leaflet.tsx
@@ -1,0 +1,190 @@
+"use client";
+
+import "leaflet/dist/leaflet.css";
+import "leaflet-draw/dist/leaflet.draw.css";
+import "leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.css";
+
+import { MapContainer } from "react-leaflet/MapContainer";
+import { TileLayer } from "react-leaflet/TileLayer";
+import { LayersControl } from "react-leaflet/LayersControl";
+import { Polygon } from "react-leaflet/Polygon";
+import { Popup } from "react-leaflet/Popup";
+import { useEffect, useState } from "react";
+import * as L from "leaflet";
+import "leaflet-draw";
+
+import DrawControl from "./Draw Control";
+import MapUpdater from "./Map Updater";
+
+import { getFarms } from "@/app/actions/actions";
+
+interface Farm {
+  id: number;
+  name: string;
+  geometry: string;
+}
+
+const MapLeaflet = ({
+  lat = -1.286389,
+  long = 36.817223,
+  height = 350,
+}: {
+  lat: number;
+  long: number;
+  height?: number;
+}) => {
+  // Map initial center coordinates
+  const position: [number, number] = [lat, long];
+
+  const [mounted, setMounted] = useState(false);
+  const [farms, setFarms] = useState<Farm[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    setMounted(true);
+    // Fix for default Leaflet icons when using Webpack/bundlers
+    L.Icon.Default.mergeOptions({
+      iconRetinaUrl: "leaflet/dist/images/marker-icon-2x.png",
+      iconUrl: "leaflet/dist/images/marker-icon.png",
+      shadowUrl: "leaflet/dist/images/marker-shadow.png",
+    });
+
+    // Async function to fetch farms
+    const fetchFarms = async () => {
+      try {
+        setLoading(true);
+        setError(null); // Clear previous errors
+        const fetchedFarms = await getFarms();
+        setFarms(fetchedFarms);
+      } catch (err: any) {
+        console.error("Error fetching farms:", err);
+        setError("Failed to load farms."); // Set an error message
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    // Call the fetch function when the component mounts
+    fetchFarms();
+  }, []); // Empty dependency array ensures this runs only once on mount
+
+  if (!mounted) return null; // Don't render map until after mount to prevent hydration issues
+
+  // Optional: Render loading or error state if needed
+  if (loading) {
+    return (
+      <div
+        style={{ height: `${height}px`, width: "100%" }}
+        className="flex items-center justify-center"
+      >
+        Loading Farms...
+      </div>
+    );
+  }
+
+  if (error) {
+    return (
+      <div
+        style={{ height: `${height}px`, width: "100%" }}
+        className="flex items-center justify-center text-red-500"
+      >
+        Error: {error}
+      </div>
+    );
+  }
+
+  return (
+    <div
+      style={{ height: `${height}px`, width: "100%" }}
+      className={`relative z-0`}
+    >
+      <MapContainer
+        center={position}
+        zoom={13}
+        scrollWheelZoom={true}
+        style={{ height: "100%", width: "100%" }}
+      >
+        {/* MapUpdater will change the view if lat/long props change */}
+        <MapUpdater lat={lat} long={long} />
+
+        {/* Base Layers Control */}
+        <LayersControl position="topright">
+          {/* OpenStreetMap */}
+          <LayersControl.BaseLayer checked name="OpenStreetMap">
+            <TileLayer
+              attribution="&copy; OpenStreetMap contributors"
+              url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+            />
+          </LayersControl.BaseLayer>
+
+          {/* Esri Satellite */}
+          <LayersControl.BaseLayer name="Satellite View (Esri)">
+            <TileLayer
+              attribution="Tiles &copy; Esri &mdash; Source: Esri"
+              url="https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}"
+            />
+          </LayersControl.BaseLayer>
+        </LayersControl>
+
+        {/* Render the fetched farms */}
+        {farms.map((farm) => {
+          try {
+            // Parse the GeoJSON string
+            const geoJson = JSON.parse(farm.geometry);
+
+            // Check if the geometry is a Polygon and has coordinates
+            if (
+              geoJson.type === "Polygon" &&
+              geoJson.coordinates &&
+              geoJson.coordinates.length > 0
+            ) {
+              // ST_AsGeoJSON returns coordinates in [longitude, latitude] order.
+              // react-leaflet's Polygon expects positions in [latitude, longitude] order.
+              // The coordinates property for a Polygon is an array of rings.
+              // We are using the first ring (index 0) which is the exterior ring.
+              const polygonPositions: [number, number][] =
+                geoJson.coordinates[0].map((coord: [number, number]) => [
+                  coord[1],
+                  coord[0],
+                ]);
+
+              // Basic path options for the polygon (customize colors/styles as needed)
+              const polygonOptions = {
+                color: "green",
+                fillColor: "#90ee90",
+                fillOpacity: 0.5,
+              };
+
+              return (
+                // Use the farm ID as the key for list rendering
+                <Polygon
+                  key={farm.id}
+                  positions={polygonPositions}
+                  pathOptions={polygonOptions}
+                >
+                  {/* Add a popup to display the farm name */}
+                  <Popup>{farm.name}</Popup>
+                </Polygon>
+              );
+            }
+            // If geometry is not a valid Polygon or has no coordinates, don't render
+            return null;
+          } catch (parseError) {
+            // Log errors for specific farms if GeoJSON parsing fails
+            console.error(
+              `Error parsing GeoJSON for farm ${farm.id}:`,
+              parseError
+            );
+            return null; // Don't render this farm
+          }
+        })}
+
+        {/* Draw Control for drawing new features (if needed) */}
+        <DrawControl />
+      </MapContainer>
+    </div>
+  );
+};
+
+export default MapLeaflet;
